Verified Buttons! Awaiting system test.
Pointed to new library inside repo, bugfix pending for error message "ignoring source code..."

+++ Arduino/libraries/CAN_Serializer/Constants.cpp
@@ -1,2 +1,2 @@
 #include <Constants.h>
-float peripheralStates[NUMSTATES] = { 0 };
+uint8_t peripheralStates[NUMSTATES] = { 0 };
